+++ packages/core-js/full/promise/all-keyed.js
@@ -1,0 +1,4 @@
+'use strict';
+var parent = require('../../actual/promise/all-keyed');
+
+module.exports = parent;