+++ packages/core-js/stage/3.js
@@ -1,6 +1,7 @@
 'use strict';
 var parent = require('./4');
 
+require('../proposals/await-dictionary');
 require('../proposals/decorator-metadata-v2');
 require('../proposals/iterator-includes');
 require('../proposals/iterator-join');