+++ tests/eslint/eslint.config.js
@@ -1364,6 +1364,10 @@ const forbidCompletelyNonExistentBuiltIns = {
     'iterateKeys',
     'iterateValues',
   ] }],
+  'es/no-nonstandard-promise-properties': [ERROR, { allow: [
+    'allKeyed',
+    'allSettledKeyed',
+  ] }],
   'es/no-nonstandard-reflect-properties': [ERROR, { allow: [
     // TODO: drop from `core-js@4`
     'defineMetadata',