+++ CHANGELOG.md
@@ -19,6 +19,10 @@
 - Added [`Iterator` join stage 3 proposal](https://github.com/tc39/proposal-iterator-join):
   - Added built-in:
     - `Iterator.prototype.join`
+- Added [Await dictionary of Promises stage 3 proposal](https://github.com/tc39/proposal-await-dictionary):
+  - Added built-ins:
+    - `Promise.allKeyed`
+    - `Promise.allSettledKeyed`
 - Throw a `RangeError` on finite unsafe integer `limit` in `Iterator.prototype.{ drop, take }`, following [tc39/ecma262/#3776](https://github.com/tc39/ecma262/pull/3776)
 - Deno 2.9+ [*replaces* `Object.prototype.__proto__`](https://github.com/denoland/deno/pull/35192) instead of removing it, so the feature detection updated
 - Ensure opaque paths always roundtrip in URL polyfill (still without adding to feature detection), [whatwg/url#844](https://github.com/whatwg/url/pull/844)